+++ WorldServer/Entities/CCheckSumCharacAccountTrunkData.cpp
@@ -108,7 +108,7 @@ bool CCheckSumCharacAccountTrunkData::Update(CRFWorldDatabase *pkDB)
   return true;
 }
 
-__int64 CCheckSumCharacAccountTrunkData::Load(CRFWorldDatabase *pkDB, CCheckSumCharacAccountTrunkData *kSrcValue)
+int CCheckSumCharacAccountTrunkData::Load(CRFWorldDatabase *pkDB, CCheckSumCharacAccountTrunkData *kSrcValue)
 {
   if (!m_dwSerial || m_dwSerial == static_cast<unsigned int>(-1) || m_dwAccountSerial == static_cast<unsigned int>(-1)
       || !pkDB)
@@ -116,7 +116,7 @@ __int64 CCheckSumCharacAccountTrunkData::Load(CRFWorldDatabase *pkDB, CCheckSumC
     return 0;
   }
 
-  unsigned int resultFlag = 0;
+  int resultFlag = 0;
   unsigned __int8 ret = pkDB->Select_NpcData(m_dwSerial, m_dwValues);
   if (ret == 1)
   {
@@ -150,7 +150,7 @@ __int64 CCheckSumCharacAccountTrunkData::Load(CRFWorldDatabase *pkDB, CCheckSumC
   return resultFlag;
 }
 
-__int64 CCheckSumCharacAccountTrunkData::CheckDiff(
+int CCheckSumCharacAccountTrunkData::CheckDiff(
   CRFWorldDatabase *pkDB,
   char *wszName,
   CCheckSumCharacAccountTrunkData *kSrcValue)
